Finishing with OOP

+++ src/Node.java
@@ -1,25 +1,24 @@
 
 public class Node implements Comparable<Node> {
-	static int createdNodes=0; 
+	static long createdNodes = 0;
 	Node parent;
 	String action;
 	int depth;
-	int cost;
+	double cost;
 	EnvironmentState state;
 
-	public Node(Node parent, String action, int depth, int cost, EnvironmentState state) {
-		this.parent=parent;
-		this.action=action;
-		this.depth=depth;
-		this.cost=cost;
-		this.state=state;
-		createdNodes++ ; 
+	public Node(Node parent, String action, int depth, double cost, EnvironmentState state) {
+		this.parent = parent;
+		this.action = action;
+		this.depth = depth;
+		this.cost = cost;
+		this.state = state;
+		createdNodes++;
 	}
 
 	@Override
 	public int compareTo(Node o) {
-		return o.state.compareTo(this.state) ; 
+		return o.state.compareTo(this.state);
 	}
-	
-	
+
 }
